requirements.txt
python-telegram-bot>=20.0
Flask>=2.0
requests>=2.0
gunicorn>=20.0
